Move OPL3, MIDIPlay, and Tester classes into own file
